// backend/src/common/types/generated/kube/models/types/v1Namespace.dto.ts
/**
 * Kubernetes
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: release-1.28
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { V1NamespaceSpec } from './v1NamespaceSpec.dto';
import { V1NamespaceStatus } from './v1NamespaceStatus.dto';
import { V1ObjectMeta } from './v1ObjectMeta.dto';
/**
 * Namespace provides a scope for Names. Use of multiple namespaces is optional.
 */
export class V1Namespace {
  /**
   * APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
   */
  'apiVersion'?: string;
  /**
   * Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
   */
  'kind'?: string;
  'metadata'?: V1ObjectMeta;
  'spec'?: V1NamespaceSpec;
  'status'?: V1NamespaceStatus;
}